App News compatible with new Params architecture

+++ apps/news/admin/main.php
@@ -23,38 +23,36 @@ class NewsAdminController extends WController {
 	}
 	
 	/**
-	 * Get the Id given in URL
-	 */
-	private function getId() {
-		$args = WRoute::getArgs();
-		if (empty($args[1])) {
-			return null;
-		} else {
-			list ($id) = explode('-', $args[1]);
-			return intval($id);
-		}
-	}
-	
-	/**
 	 * Handle News Listing action
 	 */
-	protected function listing() {
+	protected function listing(array $params) {
 		$n = 30; // Number of news per page
+		$sort_by = 'news_date';
+		$sens = 'ASC';
+		$page = 1;
 		
 		// Sorting criterias given by URL
-		$criterias = $this->getOption(0);
-		$count = sscanf(str_replace('-', ' ', $criterias), '%s %s %d', $sortBy, $sens, $page);
-		if (!isset($this->model->news_data_model['toDB'][$sortBy])) {
-			$sortBy = 'news_date';
-		}
-		if (empty($page) || $page <= 0) {
-			$page = 1;
+		if (isset($params[0])) {
+			$criterias = $params[0];
+			$count = sscanf(str_replace('-', ' ', $criterias), '%s %s %d', $sort_by_crit, $sens_crit, $page_crit);
+			
+			if (isset($this->model->news_data_model['toDB'][$sort_by_crit])) {
+				$sort_by = $sort_by_crit;
+			}
+			
+			if ($sens_crit == 'DESC') {
+				$sens = 'DESC';
+			}
+			
+			if ($page_crit > 1) {
+				$page = $page_crit;
+			}
 		}
 		
 		// SortingHelper
 		$orderingFields = array('news_id', 'news_title', 'news_author', 'news_date', 'news_views');
 		$sortingHelper = WHelper::load('SortingHelper', array($orderingFields, 'news_date', 'DESC'));
-		$sorting = $sortingHelper->findSorting($sortBy, $sens);
+		$sorting = $sortingHelper->findSorting($sort_by, $sens);
 		
 		return array(
 			'data' => $this->model->getNewsList(($page-1)*$n, $n, $sorting[0], $sorting[1] == 'ASC'),
@@ -69,7 +67,9 @@ class NewsAdminController extends WController {
 	 * - Handles Add action
 	 * - Prepares News form
 	 */
-	protected function news_form($news_id = null) {
+	protected function news_form(array $params) {
+		$news_id = isset($params[0]) ? intval($params[0]) : null;
+		
 		if (!empty($_POST)) {
 			$data = WRequest::getAssoc(array('news_author', 'news_keywords', 'news_title', 'news_url', 'news_content', 'news_cats'));
 			$errors = array();
@@ -175,15 +175,15 @@ class NewsAdminController extends WController {
 	/**
 	 * Handles Edit action
 	 */
-	protected function edit() {
-		$news_id = $this->getId();
+	protected function edit($params) {
+		$news_id = isset($params[0]) ? intval($params[0]) : -1;
 		
 		// Check whether this news exist
 		if (empty($news_id) || !$this->model->validExistingNewsId($news_id)) {
 			WNote::error('article_not_found', WLang::get('article_not_found', $news_id));
 			header('Location: '.WRoute::getDir().'/admin/news/');
 		} else {
-			return $this->news_form($news_id);
+			return $this->news_form(array($news_id));
 		}
 	}
 	
@@ -192,12 +192,13 @@ class NewsAdminController extends WController {
 	 * 
 	 * @todo Handle properly the article_not_found case with Bootstrap
 	 */
-	protected function news_delete() {
-		$news_id = $this->getId();
-		if ($this->model->validExistingNewsId($news_id)) {
+	protected function news_delete($params) {
+		$news_id = isset($params[0]) ? intval($params[0]) : -1;
+		
+		if ($news_id != -1 && $this->model->validExistingNewsId($news_id)) {
 			$data = $this->model->getNews($news_id);
 			
-			if ($this->inOptions('confirm')) {
+			if (in_array('confirm', $params)) {
 				$this->model->removeCatsFromNews($news_id);
 				$this->model->deleteNews($news_id);
 				WNote::success('article_deleted', WLang::get('article_deleted', $data['news_title']));
@@ -213,14 +214,7 @@ class NewsAdminController extends WController {
 	/**
 	 * Handles News categories_manager action
 	 */
-	protected function categories_manager() {
-		// Sorting criterias given by URL
-		$criterias = $this->getOption(0);
-		$count = sscanf(str_replace('-', ' ', $criterias), '%s %s', $sortBy, $sens);
-		if (!isset($this->model->cats_data_model['toDB'][$sortBy])) {
-			$sortBy = 'news_cat_name';
-		}
-		
+	protected function categories_manager($params) {
 		// Data was sent by form
 		if (!empty($_POST)) {
 			$data = WRequest::getAssoc(array('news_cat_id', 'news_cat_name', 'news_cat_shortname', 'news_cat_parent'));
@@ -269,10 +263,20 @@ class NewsAdminController extends WController {
 			}
 		}
 		
+		// Sorting criterias given by URL
+		$sort_by = 'news_cat_name';
+		$sens = 'ASC';
+		if (!empty($params[0])) {
+			$count = sscanf(str_replace('-', ' ', $params[0]), '%s %s', $sort_by_crit, $sens);
+			if (isset($this->model->cats_data_model['toDB'][$sort_by_crit])) {
+				$sort_by = $sort_by_crit;
+			}
+		}
+		
 		// AdminStyle Helper
 		$orderingFields = array('news_cat_name', 'news_cat_shortname');
 		$adminStyle = WHelper::load('SortingHelper', array($orderingFields, 'news_cat_name', 'ASC'));
-		$sorting = $adminStyle->findSorting($sortBy, $sens);
+		$sorting = $adminStyle->findSorting($sort_by, $sens);
 		
 		return array(
 			'data' => $this->model->getCatsList($sorting[0], $sorting[1]),
@@ -286,17 +290,18 @@ class NewsAdminController extends WController {
 	 * 
 	 * @todo Handle properly the cat_not_found case with Bootstrap
 	 */
-	protected function category_delete() {
-		$cat_id = $this->getId();
-		if ($this->model->validExistingCatId($cat_id)) {
-			if ($this->inOptions('confirm')) {
+	protected function category_delete($params) {
+		$cat_id = isset($params[0]) ? intval($params[0]) : -1;
+		
+		if ($cat_id != -1 && $this->model->validExistingCatId($cat_id)) {
+			if (in_array('confirm', $params)) {
 				$this->model->removeCatsFromNews($cat_id);
 				$this->model->unlinkChildrenOfParentCat($cat_id);
 				$this->model->deleteCat($cat_id);
 				WNote::success('category_deleted', WLang::get('category_deleted'));
 				$this->view->setHeader('Location', WRoute::getDir() . '/admin/news/categories_manager/');
 			}
-			return $cat_id;
+			return array('cat_id' => $cat_id);
 		} else {
 			WNote::error('category_not_found', WLang::get('category_not_found'));
 		}
